# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Doctor Appointment API",
        "description": "API documentation for Doctor Appointment application",
        "contact": {
            "email": "support@blogapi.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/doctors/available": {
            "get": {
                "tags": [
                    "Doctors"
                ],
                "summary": "Get available doctors with time slots",
                "description": "Returns list of doctors. When 'date' parameter is provided, includes available time slots for that specific date.",
                "operationId": "aa6d7f43f74db6a1a6baf75494aec75f",
                "parameters": [
                    {
                        "name": "specialization",
                        "in": "query",
                        "description": "Filter doctors by specialization",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Cardiology"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Date to check availability (YYYY-MM-DD). When provided, response includes available_slots array.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-11-28"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Available doctors retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/SuccessResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Dr. John Smith"
                                                            },
                                                            "specialization": {
                                                                "type": "string",
                                                                "example": "Cardiology"
                                                            },
                                                            "email": {
                                                                "type": "string",
                                                                "example": "john.smith@hospital.com"
                                                            },
                                                            "phone": {
                                                                "type": "string",
                                                                "example": "+1234567890"
                                                            },
                                                            "available_slots": {
                                                                "description": "Available time slots (only included when date parameter is provided)",
                                                                "type": "array",
                                                                "items": {
                                                                    "type": "string",
                                                                    "example": "09:00"
                                                                }
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/doctors/{doctor}": {
            "get": {
                "tags": [
                    "Doctors"
                ],
                "summary": "Get specific doctor details",
                "operationId": "d23b10111cb604261fe8961d0e96cd9f",
                "parameters": [
                    {
                        "name": "doctor",
                        "in": "path",
                        "description": "Doctor ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Date to check availability (YYYY-MM-DD). When provided, response includes available_slots array.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-11-28"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Doctor retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/SuccessResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Dr. John Smith"
                                                        },
                                                        "specialization": {
                                                            "type": "string",
                                                            "example": "Cardiology"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "john.smith@hospital.com"
                                                        },
                                                        "phone": {
                                                            "type": "string",
                                                            "example": "+1234567890"
                                                        },
                                                        "available_slots": {
                                                            "description": "Available time slots (only included when date parameter is provided)",
                                                            "type": "array",
                                                            "items": {
                                                                "type": "string",
                                                                "example": "09:00"
                                                            }
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Doctor not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Doctor]"
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/doctors/{doctor}/appointments": {
            "get": {
                "tags": [
                    "Doctors"
                ],
                "summary": "Get all appointments for a doctor",
                "operationId": "225b08de76af861be535138f9ac04acb",
                "parameters": [
                    {
                        "name": "doctor",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "confirmed",
                                "cancelled",
                                "completed"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Appointments retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/SuccessResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Appointment"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 401
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Doctor not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Doctor]"
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Doctor": {
                "required": [
                    "id",
                    "name",
                    "specialization",
                    "email"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Dr. John Smith"
                    },
                    "specialization": {
                        "type": "string",
                        "example": "Cardiology"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "doctor@example.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+1234567890"
                    }
                },
                "type": "object"
            },
            "Patient": {
                "required": [
                    "id",
                    "name",
                    "email"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Jane Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "patient@example.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+1234567890"
                    },
                    "date_of_birth": {
                        "type": "string",
                        "format": "date",
                        "example": "1990-01-15"
                    }
                },
                "type": "object"
            },
            "Appointment": {
                "description": "Appointment object. updated_at is only included when status changes.",
                "required": [
                    "id",
                    "appointment_date",
                    "status"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "appointment_date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-12-01 10:00:00"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "confirmed",
                            "cancelled",
                            "completed"
                        ],
                        "example": "pending"
                    },
                    "notes": {
                        "type": "string",
                        "example": "Follow-up consultation"
                    },
                    "updated_at": {
                        "description": "Only present when status has been changed",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-27 18:45:30"
                    },
                    "patient": {
                        "$ref": "#/components/schemas/Patient"
                    },
                    "doctor": {
                        "$ref": "#/components/schemas/Doctor"
                    }
                },
                "type": "object"
            },
...
            "SuccessResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Operation successful"
                    },
                    "statusCode": {
                        "type": "integer",
                        "example": 200
                    },
                    "status": {
                        "type": "string",
                        "example": "OK"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT Bearer token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}